api/pyproject.toml
[tool.poetry]
name = "manifoods-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
alembic = "^1.7.5"
psycopg2-binary = "^2.9.2"
databases = {extras = ["postgresql"], version = "^0.4.3"}
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
uvicorn = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
